Include the contents of @rescript/runtime in the artifact list
