+++ packages/artifacts.json
@@ -1,0 +1,27 @@
+{
+  "rescript": [
+    "CHANGELOG.md",
+    "COPYING",
+    "COPYING.LESSER",
+    "CREDITS.md",
+    "LICENSE",
+    "README.md",
+    "cli/bsc.js",
+    "cli/bstracing.js",
+    "cli/common/args.js",
+    "cli/common/bins.js",
+    "cli/common/bsb.js",
+    "cli/common/minisocket.js",
+    "cli/rescript-legacy.js",
+    "cli/rescript-legacy/dump.js",
+    "cli/rescript-legacy/format.js",
+    "cli/rescript-tools.js",
+    "cli/rescript.js",
+    "docs/docson/build-schema.json",
+    "ninja/COPYING",
+    "package.json"
+  ],
+  "@rescript/runtime": [
+    "package.json"
+  ]
+}
