--- packages/artifacts.txt
@@ -1,20 +1,0 @@
-CHANGELOG.md
-COPYING
-COPYING.LESSER
-CREDITS.md
-LICENSE
-README.md
-cli/bsc.js
-cli/bstracing.js
-cli/common/args.js
-cli/common/bins.js
-cli/common/bsb.js
-cli/common/minisocket.js
-cli/rescript-legacy.js
-cli/rescript-legacy/dump.js
-cli/rescript-legacy/format.js
-cli/rescript-tools.js
-cli/rescript.js
-docs/docson/build-schema.json
-ninja/COPYING
-package.json
